apps/api/pyproject.toml
[project]
name = "continuum-api"
version = "0.1.0"
description = "Continuum Knowledge Management API"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.18.3",
    "neo4j>=5.27.0",
    "redis>=5.2.1",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.10.4",
    "pydantic-settings>=2.7.1",
    "websockets>=14.1",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "bcrypt>=4.2.1",
    # SEC-013: python-jose with cryptography for JWT handling
    # Using version constraint to avoid known vulnerabilities in older versions
    "python-jose[cryptography]>=3.3.0,<4.0",
    "rapidfuzz>=3.0.0",
    "watchdog>=4.0.0",
    "prometheus-client>=0.20.0",
    # SEC-013: Explicit cryptography version for security updates
    "cryptography>=42.0.0",
    # SEC-013: Starlette pinned for security (dependency of FastAPI)
    "starlette>=0.52.1",
    # Hackathon: Amazon Bedrock (Claude) provider
    "boto3>=1.35.0",
    # Hackathon: Datadog LLM Observability
    "ddtrace>=2.0.0",
    # Strands Agents SDK (Bedrock + OpenAI-compatible providers)
    "strands-agents[openai]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.6",
    "mypy>=1.14.1",
    # SEC-013: Security audit tools
    "pip-audit>=2.7.0",
    "safety>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "N818"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: marks tests as unit tests (fast, no I/O)",
    "integration: marks tests as integration tests (require services)",
    "e2e: marks tests as end-to-end tests (require full stack)",
    "slow: marks tests as slow running",
]

[tool.coverage.run]
source = ["services", "routers", "agents", "middleware"]
omit = ["tests/*", ".venv/*"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
